Fix: Config file use before update
- Add get_config as a public config loading function
